--- src/app/(provider)/(loggedInOnly)/profiles/page.tsx
@@ -1,5 +1,0 @@
-function ProfilePage() {
-  return <div>ProfilePage</div>;
-}
-
-export default ProfilePage;
