+++ src/app/(provider)/(root)/(loggedInOnly)/profiles/page.tsx
@@ -1,0 +1,7 @@
+import Page from "@/components/Page/Page";
+
+function ProfilePage() {
+  return <Page width="lg">ProfilePage</Page>;
+}
+
+export default ProfilePage;
